convert to float proactively

+++ openproblems/data/sample.py
@@ -20,9 +20,8 @@ def load_sample_data(test=True):
     cells = pd.read_csv(rna_cells_url, low_memory=False, index_col=0).iloc[:200]
 
     rna_data = scipy.sparse.csr_matrix(
-        np.random.poisson(0.3, (cells.shape[0], genes.shape[0]))
+        np.random.poisson(0.3, (cells.shape[0], genes.shape[0])).astype(np.float64)
     )
 
     adata = anndata.AnnData(rna_data, obs=cells, var=genes)
-    adata.X = adata.X.astype(np.float64)
     return adata
